ML/test/requirements.txt
Flask
pandas
tensorflow
numpy
matplotlib.pyplot 